ADD first gemma version

+++ ms-gemma/training/train_lora.py
@@ -40,7 +40,7 @@ def train():
     bnb_config = BitsAndBytesConfig(
         load_in_4bit=True,
         bnb_4bit_quant_type="nf4",
-        bnb_4bit_compute_dtype=torch.float16,
+        bnb_4bit_compute_dtype=torch.bfloat16,
         bnb_4bit_use_double_quant=True,
     )
 
@@ -90,7 +90,7 @@ def train():
         gradient_accumulation_steps=4,
         learning_rate=2e-4,
         logging_steps=10,
-        fp16=True,
+        bf16=True,
         optim="paged_adamw_32bit",
         save_strategy="epoch",
         warmup_steps=10,
